nuevas funciones en pago y reservas

+++ js/helpers/cargarReserva.js
@@ -4,81 +4,39 @@ export const cargarReserva = (props) => {
 
     const socioActual = JSON.parse(localStorage.getItem('socioDatos'));
 
-    // Las siguientes 3 funciones permiten determinar los horarios disponibles para cada lugar
-    const horariosCancha = (fecha) => {
-        let horariosCancha = [];
-        fetch("js/helpers/horarios.json")
-            .then(res => {
-                if(!res.ok){
-                    throw new Error('Error al cargar los horarios');
-                }
-                return res.json();
-            })
-            .then(data => {
-                data.forEach(horario => {
-                    if (horario.lugar === 'Cancha') {
-                        horario.disponibilidad.forEach(hora => {
-                            if (hora.fecha === fecha && hora.disponibilidad === true) {
-                                horariosCancha.push(hora.hora);
-                            }
-                        });
+    // Las siguientes 2 funciones permiten determinar los horarios disponibles para cada lugar
+
+    const obtenerHorarios = async () => {
+        try {
+            const res = await fetch("js/helpers/horarios.json");
+            if (!res.ok) {
+                throw new Error('Error al cargar los horarios');
+            }
+            return await res.json();
+        } catch (error) {
+            console.log("El error es: ", error);
+        }
+    };
+
+    const horariosLugar = async (fecha, lugar) => {
+        const data = await obtenerHorarios();
+        let horariosLugar = [];
+        data.forEach(horario => {
+            if (horario.lugar === lugar) {
+                horario.disponibilidad.forEach(hora => {
+                    if (hora.fecha === fecha && hora.disponibilidad === true) {
+                        horariosLugar.push(hora.hora);
                     }
                 });
-            })
-            .catch(err => console.log("El error es: " , err));
-        return horariosCancha;
-    }
-    const horariosPiscina = (fecha) => {
-        let horariosPiscina = [];
-        fetch("js/helpers/horarios.json")
-            .then(res => {
-                if(!res.ok){
-                    throw new Error('Error al cargar los horarios');
-                }
-                return res.json();
-            })
-            .then(data => {
-                data.forEach(horario => {
-                    if (horario.lugar === 'Piscina') {
-                        horario.disponibilidad.forEach(hora => {
-                            if (hora.fecha === fecha && hora.disponibilidad === true) {
-                                horariosPiscina.push(hora.hora);
-                            }
-                        });
-                    }
-                });
-            })
-            .catch(err => console.log("El error es: " , err));
-        return horariosPiscina;
-    }
-    const horariosSalon = (fecha) => {
-        let horariosSalon = [];
-        fetch("js/helpers/horarios.json")
-            .then(res => {
-                if(!res.ok){
-                    throw new Error('Error al cargar los horarios');
-                }
-                return res.json();
-            })
-            .then(data => {
-                data.forEach(horario => {
-                    if (horario.lugar === 'Salón') {
-                        horario.disponibilidad.forEach(hora => {
-                            if (hora.fecha === fecha && hora.disponibilidad === true) {
-                                horariosSalon.push(hora.hora);
-                            }
-                        });
-                    }
-                });
-            })
-            .catch(err => console.log("El error es: " , err));
-        return horariosSalon;
-    }
+            }
+        });
+        return horariosLugar;
+    };
 
     // El siguiente evento permiten capturar la eleccion de fecha y a partir de esta los horarios disponibles para cada lugar
     const inputFecha = document.querySelectorAll('.fecha');
     inputFecha.forEach(input => {
-        input.addEventListener('change', (e) => {
+        input.addEventListener('change', async (e) => {
             e.preventDefault();
             let fecha = input.value;
             let lugar = input.parentElement.querySelector('.lugar').innerHTML;
@@ -87,7 +45,8 @@ export const cargarReserva = (props) => {
                 let selectCancha = input.parentElement.querySelector('#selectCancha');
                 selectCancha.innerHTML = '';
 
-                horariosCancha(fecha).forEach(hora => {
+                const horarios = await horariosLugar(fecha, "Cancha");
+                horarios.forEach(hora => {
                     let option = document.createElement('option');
                     option.innerHTML = hora;
                     selectCancha.appendChild(option);
@@ -96,7 +55,8 @@ export const cargarReserva = (props) => {
                 let selectPiscina = input.parentElement.querySelector('#selectPiscina');
                 selectPiscina.innerHTML = '';
 
-                horariosPiscina(fecha).forEach(hora => {
+                const horarios = await horariosLugar(fecha, "Piscina");
+                horarios.forEach(hora => {
                     let option = document.createElement('option');
                     option.innerHTML = hora;
                     selectPiscina.appendChild(option);
@@ -105,7 +65,8 @@ export const cargarReserva = (props) => {
                 let selectSalon = input.parentElement.querySelector('#selectSalon');
                 selectSalon.innerHTML = '';
 
-                horariosSalon(fecha).forEach(hora => {
+                const horarios = await horariosLugar(fecha, "Salón");
+                horarios.forEach(hora => {
                     let option = document.createElement('option');
                     option.innerHTML = hora;
                     selectSalon.appendChild(option);
@@ -134,28 +95,16 @@ export const cargarReserva = (props) => {
         });
     });
 
-    // En este evento teniendo en cuenta los socios y sus reservas modifica el estado del boton de reserva por cada uno de los formularios de la paginna actual
-    socios.forEach(socio => {
-        if (socio.usuario !== socioActual.usuario) {
-            socio.reservas.forEach(reserva => {
-                const formReserva = document.querySelectorAll('.form-reserva');
-                formReserva.forEach(form => {
-                    if (form.querySelector('.lugar').innerHTML === reserva.lugar && reserva.estado === 'Reservado') {
-                        form.querySelector('.fecha').value = reserva.fecha;
-                        form.querySelector('.hora').value = reserva.hora;
-
-                        let btnReserva = form.querySelector('.btn-reserva');
-                        btnReserva.innerHTML = "No disponible";
-                        btnReserva.disabled = true;
-                        form.querySelector('.fecha').disabled = true;
-                        form.querySelector('.hora').disabled = true;
-                    }
-                });
+    // La siguiente funcion permite verificar si existe una reserva para el lugar, fecha y hora seleccionados
+    const existeReserva = (lugar, fecha, hora) => {
+        return socios.some(socio => {
+            return socio.reservas.some(reserva => {
+                return reserva.lugar === lugar && reserva.fecha === fecha && reserva.hora === hora;
             });
-        }
-    });
+        });
+    }
 
-    // En este evento permite capturar el bonton de la reserva guardando la misma en el local storage y modificando el estado del boton.
+    // En este evento permite capturar el boton de la reserva guardando la misma en el local storage y modificando el estado del boton.
     const btnReserva = document.querySelectorAll('.btn-reserva');
     btnReserva.forEach(btn => {
         btn.addEventListener('click', (e) => {
@@ -165,27 +114,36 @@ export const cargarReserva = (props) => {
             let hora = btn.parentElement.querySelector('.hora').value;
             let estado = 'Reservado';
 
+            console.log(existeReserva(lugar, fecha, hora));
+
             if (fecha !== '' && hora !== '') {
-                let reserva = {
-                    lugar,
-                    fecha,
-                    hora,
-                    estado
+                if (!existeReserva(lugar, fecha, hora)) {
+                    let reserva = {
+                        lugar,
+                        fecha,
+                        hora,
+                        estado
+                    }
+                    socioActual.reservas.push(reserva);
+
+                    localStorage.setItem('socioDatos', JSON.stringify(socioActual));
+
+                    btn.innerHTML = 'Reservado';
+                    btn.parentElement.querySelector('.fecha').disabled = true;
+                    btn.parentElement.querySelector('.hora').disabled = true;
+                    btn.disabled = true;
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Reserva realizada',
+                        text: `Se realizo la reserva de ${lugar} para el ${fecha} a las ${hora}`,
+                    })
+                } else {
+                    Swal.fire({
+                        icon: 'info',
+                        title: 'Reserva existente',
+                        text: 'Ya existe una reserva para el lugar, fecha y hora seleccionados'
+                    })
                 }
-
-                socioActual.reservas.push(reserva);
-
-                localStorage.setItem('socioDatos', JSON.stringify(socioActual));
-
-                btn.innerHTML = 'Reservado';
-                btn.parentElement.querySelector('.fecha').disabled = true;
-                btn.parentElement.querySelector('.hora').disabled = true;
-                btn.disabled = true;
-                Swal.fire({
-                    icon: 'success',
-                    title: 'Reserva realizada',
-                    text: `Se realizo la reserva de ${lugar} para el ${fecha} a las ${hora}`,
-                })
             } else {
                 Swal.fire({
                     icon: 'error',
